Extract EphemeristsSelectCard to its own file, byte-for-byte (#2323)
The pure move, so the repaint that follows has nothing structural in it.
`EphemeristsSelectCard` leaves `FactionSelectCard.tsx` for
`components/selectCard/EphemeristsSelectCard.tsx`, matching how
`components/taskCard/` is laid out and following the shape #2322 set. The
manifest row is unchanged in shape — `lazyArchetype(() => import(...))` now
points at the new module and drops the `.then((m) => ({ default:
m.EphemeristsSelectCard }))` unwrap, because the new file default-exports it
the way every task card does.

PROOF THE DIRECTORY IS UNCHANGED. `renderToStaticMarkup` of
`<FactionSelectCard faction="ephemerists" />` across all six state x members
combinations is byte-identical before and after this commit (47,526 bytes,
`diff` clean). The probe was a throwaway; the standing guard is
`factionSelectCardDispatch.test.tsx`, green.

`factionName` leaves the dispatcher's imports with the block: the wordmark on
this tile was its last reader there.
